qr-reader-attendance-system/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
opencv-python==4.9.0.80
pyzbar==0.1.9
python-dotenv==1.0.1
Werkzeug==3.0.1
Pillow==10.2.0
qrcode==7.4.2
pandas==2.2.1
